rint() shifts 1 into the sign bit, on every negative argument
  lib/msun/src/s_rint.c:36    int32_t i0,j0,sx;
  lib/msun/src/s_rint.c:40    sx = (i0>>31)&1;
  lib/msun/src/s_rint.c:52    SET_HIGH_WORD(t,(i0&0x7fffffff)|(sx<<31));

sx is int32_t and is 0 or 1. For a negative argument it is 1, and
1 << 31 is not representable in a signed 32-bit int - C17 6.5.7p4,
undefined behaviour. s_rintf.c has the same line.

Not a corner case: rint(3) takes that branch for every negative input
with |x| < 2^52, which is most of them.

  $ gcc -fsanitize=undefined -fno-sanitize-recover=all
  runtime error: left shift of 1 by 31 places cannot be represented
                 in type 'int'

Cast to u_int32_t, which is what the surrounding code already uses and
what SET_HIGH_WORD wants anyway. The bit pattern is unchanged - 0x80000000
before and after - because the UB was in the type of the shift, not in the
value it computes, which is why this has never been seen.

  CBMC after: s_rint.c:rint PROVED, s_rintf.c:rintf PROVED
  UBSan after: clean, hi == 0x80000000 as required

== and the fourth calibration, which this run needed

--div-by-zero-check does not distinguish integer from floating-point
division, and only integer division by zero is undefined:

  $ cbmc fdiv.c --function fdiv --div-by-zero-check   /* double a / b */
  [fdiv.division-by-zero.1] division by zero in a / b: FAILURE

IEEE-754 defines x / 0.0 as an infinity with a flag raised, and lib/msun
RELIES on it - that is how log(0), logb, rsqrt and the whole catrig family
produce their infinities. Thirty-odd of this run's failures in lib/msun
were that, and not one was a defect.

Recorded in the README rather than quietly dropped, because the class is
not uniformly noise: division by zero in lib/libc/stdlib/getopt_long.c's
gcd() is integer, and is worth looking at.

That is now four calibrations, all found by looking at a first result that
was obviously wrong rather than by reading the manual:

  --conversion-check     defined (C17 6.3.1.3p2), would have "found" a bug
                         in three-line ffs()
  pointer parameters     a nondeterministic char * includes NULL, so every
                         string routine "fails" its own missing precondition
  extern pointer globals the same trap for a function taking no arguments
                         at all - getchar() reads __stdinp
  float division by zero defined by IEEE-754, and depended upon

+++ hbsd/src/lib/msun/src/s_rint.c
@@ -49,7 +49,7 @@ rint(double x)
 	        STRICT_ASSIGN(double,w,TWO52[sx]+x);
 	        t =  w-TWO52[sx];
 		GET_HIGH_WORD(i0,t);
-		SET_HIGH_WORD(t,(i0&0x7fffffff)|(sx<<31));
+		SET_HIGH_WORD(t,(i0&0x7fffffff)|((u_int32_t)sx<<31));
 	        return t;
 	    } else {
 		i = (0x000fffff)>>j0;
